new editor service
