+++ backend/internal/handler/draft.go
@@ -25,106 +25,6 @@ func NewDraftHandler(draftService *DraftService, chartService *ChartService) *Dr
 // ====================================================================================
 // Handlers
 
-func (a *DraftHandler) SeparateAudio(c *gin.Context) {
-	UUID := c.PostForm("uuid")
-	file, err := c.FormFile("audio")
-	if err != nil {
-		transport.BadRequest(c, "no audio file provided")
-		return
-	}
-	vocalsURL, instrumentalURL, err := a.draftService.SeparateAudio(c.Request.Context(), UUID, file)
-	if err != nil {
-		transport.ServiceError(c, err)
-		return
-	}
-	type Response struct {
-		VocalsURL       string `json:"vocalsUrl"`
-		InstrumentalURL string `json:"instrumentalUrl"`
-	}
-	c.JSON(http.StatusOK, Response{
-		VocalsURL:       vocalsURL,
-		InstrumentalURL: instrumentalURL,
-	})
-}
-
-func (a *DraftHandler) UploadInstrumental(c *gin.Context) {
-	UUID := c.PostForm("uuid")
-	file, err := c.FormFile("instrumental")
-	if err != nil {
-		transport.BadRequest(c, "no audio file provided")
-		return
-	}
-	audioURL, err := a.draftService.UploadInstrumental(c.Request.Context(), UUID, file)
-	if err != nil {
-		transport.ServiceError(c, err)
-		return
-	}
-	type Response struct {
-		AudioURL string `json:"audioUrl"`
-	}
-	c.JSON(http.StatusOK, Response{AudioURL: audioURL})
-}
-
-func (a *DraftHandler) UploadVocals(c *gin.Context) {
-	UUID := c.PostForm("uuid")
-	file, err := c.FormFile("vocals")
-	if err != nil {
-		transport.BadRequest(c, "no audio file provided")
-		return
-	}
-	audioURL, err := a.draftService.UploadVocals(c.Request.Context(), UUID, file)
-	if err != nil {
-		transport.ServiceError(c, err)
-		return
-	}
-	type Response struct {
-		AudioURL string `json:"audioUrl"`
-	}
-	c.JSON(http.StatusOK, Response{AudioURL: audioURL})
-}
-
-func (a *DraftHandler) UploadImage(c *gin.Context) {
-	UUID := c.PostForm("uuid")
-	file, err := c.FormFile("image")
-	if err != nil {
-		transport.BadRequest(c, "no image file provided")
-		return
-	}
-	imageURL, err := a.draftService.UploadImage(c.Request.Context(), UUID, file)
-	if err != nil {
-		transport.ServiceError(c, err)
-		return
-	}
-	type Response struct {
-		ImageURL string `json:"imageUrl"`
-	}
-	c.JSON(http.StatusOK, Response{ImageURL: imageURL})
-}
-
-func (a *DraftHandler) GenerateTimings(c *gin.Context) {
-	UUID := c.PostForm("uuid")
-	lyrics := c.PostForm("lyrics")
-	if UUID == "" {
-		transport.BadRequest(c, "missing uuid")
-		return
-	}
-	if lyrics == "" {
-		transport.BadRequest(c, "missing lyrics")
-		return
-	}
-	timings, err := a.draftService.GenerateTimings(c.Request.Context(), UUID, lyrics)
-	if err != nil {
-		transport.ServiceError(c, err)
-		return
-	}
-	type Response struct {
-		Timings []t.Timing `json:"timings"`
-	}
-	c.JSON(http.StatusOK, Response{Timings: timings})
-}
-
-// Draft CRUD
-
 func (a *DraftHandler) InitDraft(c *gin.Context) {
 	uuid, err := a.draftService.InitDraft(c.Request.Context())
 	if err != nil {
