+++ backend/internal/service/editor/errors.go
@@ -1,0 +1,15 @@
+package editor
+
+import "errors"
+
+var (
+	ErrMissingUUID               = errors.New("missing uuid")
+	ErrInvalidAudioType          = errors.New("invalid audio type")
+	ErrInvalidImageType          = errors.New("invalid image type")
+	ErrSeparationFailed          = errors.New("separation failed")
+	ErrVocalsNotGenerated        = errors.New("vocals file not generated")
+	ErrInstrumentalNotGenerated  = errors.New("instrumental file not generated")
+	ErrAlignmentFailed           = errors.New("alignment failed")
+	ErrParsingLyrics             = errors.New("failed to parse lyrics")
+	ErrPythonInterpreterNotFound = errors.New("python interpreter not found")
+)
